ggheatmap {heatmaply}	R Documentation

ggplot heatmap equivalent to heatmaply

Description

This function produces a ggplot analogue of heatmaply figures using ggarrange. This function may not always support the same set of features as , and exporting the heatmaply object with, for example, orca or heatmaply(mtcars, file = "foo.png").

Usage

ggheatmap(..., widths = NULL, heights = NULL, row_dend_left = FALSE)

Arguments

...
Passed to heatmaply

widths, heights
Relative widths and heights of plots.

row_dend_left
Logical argument controlling whether the row dendrogram is placed on the left of the plot.

Examples

ggheatmap(mtcars)

[Package heatmaply version 1.0.0 Index]