ggplot_side_color_plot {heatmaply}	R Documentation

Side color plots for heatmaps

Description

Important for creating annotation.

Usage

ggplot_side_color_plot(df, palette = NULL, scale_title = paste(type,
  "side colors"), type = c("column", "row"), text_angle = if (type ==
  "column") 0 else 90, is_colors = FALSE, fontsize, label_name = NULL)

Arguments

df
A "molten" data.frame as produced by (eg) reshape2::melt

palette
A function which can return colors to be used in the sidebar plot

scale_title
Title of the color scale. Not currently used.

type
Horizontal or vertical plot? Valid values are "column" and "row"

text_angle
the angle of the text of the rows/columns.

is_colors
Use if the values in df are valid colours and should not be mapped to a color scheme, and instead should be plotted directly.

fontsize
Font size (currently unused)

label_name
Name for the mouseover label, usually "row" or "column"

Value

A ggplot geom_tile object

[Package heatmaply version 0.15.2 Index]