normalize {heatmaply}	R Documentation

Normalization transformation (0-1)

Description

An Empirical Normalization Transformation brings data to the 0 to 1 scale by substracting the minimum and dividing by the maximum of all observations. This is similar to percentize in that it allows to compare variables of different scales, but it also keeps the shape of the distribution.

Usage

normalize(x, ...)

Arguments

x
a vector or a data.frame.

...
Currently ignored.

Value

A vector (or data.frame) after normalizing the numeric variables.

See Also

percentize

Examples

## Not run: 
x <- mtcars
x <- data.frame(x)
x$am <- factor(x$am)
x$vs <- factor(x$vs)
heatmaply(percentize(x))
heatmaply(normalize(x))

x <- data.frame(a = 1:10, b = 11:20)
x[4:6, 1:2] <- NA
normalize(x)
normalize(x[,1])

## End(Not run)

[Package heatmaply version 0.15.2 Index]